chore: create separate folder for util fns
