+++ src/utils/assert.ts
@@ -1,7 +1,3 @@
-export function roundToTwoDecimals(value: number) {
-  return Number(Math.round(parseFloat(value + "e" + 2)) + "e-" + 2);
-}
-
 export function assertNonNullish<T>(
   value: T,
   message: string,
